Restore collisions between poop and player.
After moving the poop to the EntityManager, the player lost the ability to
destroy the poop, because collision detection would stop the player from
moving over the poop.

Since the player can destroy the poop only from above, this adds a check
for collisions between the poop and the player that extends the poop's
bounding box in that direction.