Fix rounding error in positioning of the player against a tile.
In some cases, if the player is surrounded by solid tiles, the computation
done in getTopBound() returns a very slightly different result compared to
what would be produced by using y + bounds.y, because the operation order
is subtly different. This tiny difference is enough to make it look as if
the player entered the solid tile, which would make the collision detection
move the player away by a whole tile.