Align the poop impact with the edge of the tile.
The change in 4a5a49c made the poop impact appear inside of a tile if the
poop was fired from very close to the tile. The fix is to allow the poop to
go up more, so that it can align to the edge of the tile. To avoid breaking
a tile above the player, we don't allow an up movement of more that half a
tile.